Move System.Native/pal_maphardwaretype.cpp to C

+++ src/Native/Unix/System.Native/pal_interfaceaddresses.cpp
@@ -7,6 +7,7 @@
 #include "pal_maphardwaretype.h"
 #include "pal_utilities.h"
 #include "pal_safecrt.h"
+#include "pal_networking.h"
 
 #include <sys/types.h>
 #include <assert.h>
@@ -114,7 +115,7 @@ extern "C" int32_t SystemNative_EnumerateInterfaceAddresses(IPv4AddressFound onI
                 memset(&lla, 0, sizeof(LinkLayerAddressInfo));
                 lla.InterfaceIndex = interfaceIndex;
                 lla.NumAddressBytes = sall->sll_halen;
-                lla.HardwareType = MapHardwareType(sall->sll_hatype);
+                lla.HardwareType = static_cast<uint8_t>(MapHardwareType(sall->sll_hatype));
 
                 memcpy_s(&lla.AddressBytes, sizeof_member(LinkLayerAddressInfo, AddressBytes), &sall->sll_addr, sall->sll_halen);
                 onLinkLayerFound(current->ifa_name, &lla);
@@ -130,7 +131,7 @@ extern "C" int32_t SystemNative_EnumerateInterfaceAddresses(IPv4AddressFound onI
                 LinkLayerAddressInfo lla = {
                     .InterfaceIndex = interfaceIndex,
                     .NumAddressBytes = sadl->sdl_alen,
-                    .HardwareType = MapHardwareType(sadl->sdl_type),
+                    .HardwareType = static_cast<uint8_t>(MapHardwareType(sadl->sdl_type)),
                 };
 
                 memcpy_s(&lla.AddressBytes, sizeof_member(LinkLayerAddressInfo, AddressBytes), reinterpret_cast<uint8_t*>(LLADDR(sadl)), sadl->sdl_alen);
